Fix compiler error with avr-gcc 4.7 and the ntpTest example.
With avr-gcc 4.7, which seems to be the version used by installing arduino as a package on  debian/ubuntu, the ntpTest example fails to compile with errors such as:
ntpTest.ino:78:22: error: variable ‘janStr’ must be const in order to be put into read-only section by means of ‘__attribute__((progmem))’

This hasn't been seen before because the default version of avr-gcc used by arduino is 4.3, which doesn't catch this issue.  This fix to the example resolves the compilation problem with avr-gcc 4.7.  The commit was tested by running the ntpTest example with both avr-gcc 4.3 and 4.7 versions of Arduino--the example ran as expected in both cases.