Fix issue when running with Arduino and avr-gcc
With avr-gcc 4.7.2 (but curiously not with the older 4.3 version) it was necessary to add a copy constructor and assignment operator to the client class.  The default versions of these that were being generated were causing the CC3000 to become unresponsive.  Also calls to interrupt polling functions were added to fix instability found when compiled with avr-gcc 4.3.