Fix WebClient example to allow reading data before connection is closed.
See discussion in this issue for more context: https://github.com/adafruit/Adafruit_CC3000_Library/issues/31

This change makes the WebClient example use HTTP/1.1 so the connection is persistent and will not be closed before all the data can be read by the sketch.  A simple 3 second idle timeout is added to detect when the server has finished sending data and the sketch should exit.

This change was tested by running the modified web client example with the normal adafruit web page, and pages mentioned in the linked issue.