Add optional CC3K debug printer to main class constructor.
This optional parameter allows a user to specify their own Print object for receiving debug output.  The default behavior of choosing hardware serial 1 is unchanged if the parameter is not specified.  This change follows the conventions of Arduino libraries and allows a user to use the hardware serial port without running into problems with the CC3K attempting to use it too.

This change was tested by running all the examples from the CC3000 test suite (here https://github.com/ladyada/CC3000_testing ) and verifying there were no regressions.  I also ran a version of the buildtest example which specified a null value for the printer and verified no debug output was written.